pyproject.toml
[project]
name = "agentic-flow"
version = "0.35"
description = "Minimal wrapper over OpenAI Agent SDK with method chaining and ChatKit integration"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "openai-agents>=0.3.2",
    "openai-chatkit>=1.4.0,<2",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src/agentic_flow"]

[dependency-groups]
dev = [
    "black>=24.0",
    "mkdocs-material>=9.0",
    "pytest>=9.0.2",
    "pytest-asyncio>=1.3.0",
    "python-dotenv>=1.2.1",
    "ruff>=0.14.10",
]
sample = [
    "fastapi>=0.114.1",
    "uvicorn[standard]>=0.30",
    "openai>=1.40",
    "pydantic>=2.0",
    "rich>=13.0",
    "textual>=0.89.0",
]

[tool.ruff]
line-length = 100
target-version = "py312"

[tool.ruff.lint]
select = ["E", "F", "I", "UP"]
ignore = ["UP046"]  # Generic[T] syntax is clearer than Python 3.12 type params

[tool.ruff.format]
quote-style = "double"
indent-style = "space"

[tool.black]
line-length = 100
target-version = ["py312"]

[tool.pytest.ini_options]
asyncio_mode = "auto"
testpaths = ["tests"]

[tool.mypy]
ignore_missing_imports = false

[[tool.mypy.overrides]]
module = "chatkit.*"
ignore_missing_imports = true
